+++ soul-client/soul-client-dubbo/soul-client-alibaba-dubbo/src/test/java/org/dromara/soul/client/alibaba/dubbo/validation/mock/MockValidationParameter.java
@@ -28,9 +28,7 @@ import javax.validation.constraints.NotNull;
  */
 @Data
 public class MockValidationParameter {
-    /**
-     * String parameter.
-     */
+    
     @NotNull
     private String stringParameter;
