+++ soul-client/soul-client-dubbo/soul-client-alibaba-dubbo/src/test/java/org/dromara/soul/client/alibaba/dubbo/validation/mock/MockValidatorTarget.java
@@ -29,6 +29,7 @@ import java.util.Map;
  * @author David Liu
  */
 public interface MockValidatorTarget {
+    
     /**
      * mock method: method1.
      *
