+++ soul-client/soul-client-dubbo/soul-client-apache-dubbo/src/main/java/org/dromara/soul/client/apache/dubbo/validation/ApacheDubboClientValidation.java
@@ -22,7 +22,7 @@ import org.apache.dubbo.validation.Validation;
 import org.apache.dubbo.validation.Validator;
 
 /**
- * ApacheDubboClientValidation.
+ * The type Apache dubbo client validation.
  *
  * @author KevinClair
  */
